Redistribute controller IT tests to their respective specs
