+++ src/test/groovy/com/paiondata/aristotle/AbstractITSpec.groovy
@@ -17,8 +17,17 @@ package com.paiondata.aristotle
 
 import com.paiondata.aristotle.base.TestConstants
 
+import org.springframework.http.HttpStatus
+import org.testcontainers.containers.DockerComposeContainer
+import org.testcontainers.containers.wait.strategy.Wait
+import org.testcontainers.spock.Testcontainers
+
 import io.restassured.http.ContentType
 import org.junit.jupiter.api.Assertions
+
+import spock.lang.Shared
+
+import java.time.Duration
 
 import javax.validation.constraints.NotNull
 import org.junit.Assert
@@ -29,24 +38,53 @@ import spock.lang.Specification
 import java.nio.file.Files
 import java.nio.file.Paths
 
+@Testcontainers
 abstract class AbstractITSpec extends Specification {
 
-    protected static final int WS_PORT = 8080
-    private static final int OK_CODE = 200
-    private static final String USER_ENDPOINT = "/user"
-    private static final String GRAPH_ENDPOINT = "/graph"
-    private static final String NODE_ENDPOINT = "/node"
-    private static final String CREATE_UPDATE_USER_JSON = "create-update-user.json"
-    private static final String UPDATE_GRAPH_JSON = "update-graph.json"
-    private static final String GET_GRAPH_JSON = "get-graph.json"
-    private static final String DELETE_GRAPH_JSON = "delete-graph.json"
-    private static final String UPDATE_NODE_JSON = "update-node.json"
-    private static final String TEST_UIDCID = "6b47"
-    private static final String TEST_NICK_NAME = "Jame"
-    private static final String UPDATE_NICK_NAME = "Fame"
-    private static final String TEST_GRAPH_TITLE = "Rus"
-    private static final String UPDATE_GRAPH_TITLE = "Kas"
-    private static final String UPDATE_NODE_TITLE = "Los"
+    static final int WS_PORT = 8080
+    static final String USER_ENDPOINT = "/user"
+    static final String GRAPH_ENDPOINT = "/graph"
+    static final String NODE_ENDPOINT = "/node"
+    static final String CREATE_UPDATE_USER_JSON = "create-update-user.json"
+    static final String UPDATE_GRAPH_JSON = "update-graph.json"
+    static final String GET_GRAPH_JSON = "get-graph.json"
+    static final String DELETE_GRAPH_JSON = "delete-graph.json"
+    static final String UPDATE_NODE_JSON = "update-node.json"
+    static final String TEST_UIDCID = "6b47"
+    static final String TEST_NICK_NAME = "Jame"
+    static final String UPDATE_NICK_NAME = "Fame"
+    static final String TEST_GRAPH_TITLE = "Rus"
+    static final String UPDATE_GRAPH_TITLE = "Kas"
+    static final String UPDATE_NODE_TITLE = "Los"
+
+    @Shared
+    DockerComposeContainer COMPOSE = new DockerComposeContainer(new File("docker-compose.yml"))
+            .withExposedService(
+                    "web",
+                    WS_PORT,
+                    Wait.forHttp("/actuator/health").forStatusCode(HttpStatus.OK.value())
+            ).withStartupTimeout(Duration.ofMinutes(10))
+
+    def childSetupSpec() {
+        // intentionally left blank
+    }
+
+    def childCleanupSpec() {
+        // intentionally left blank
+    }
+
+    def setupSpec() {
+        RestAssured.baseURI = "http://localhost"
+        RestAssured.port = WS_PORT
+
+        childSetupSpec()
+    }
+
+    def cleanupSpec() {
+        RestAssured.reset()
+
+        childCleanupSpec()
+    }
 
     @NotNull
     protected String payload(final @NotNull String resourceName) {
@@ -88,271 +126,6 @@ abstract class AbstractITSpec extends Specification {
         }
     }
 
-    def childSetupSpec() {
-        // intentionally left blank
-    }
-
-    def childCleanupSpec() {
-        // intentionally left blank
-    }
-
-    def setupSpec() {
-        childSetupSpec()
-
-        RestAssured.baseURI = "http://localhost"
-        RestAssured.port = WS_PORT
-    }
-
-    def cleanupSpec() {
-        RestAssured.reset()
-
-        childCleanupSpec()
-    }
-
-    def "JSON API handles invalid user creation requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload(CREATE_UPDATE_USER_JSON), uidcid, nickName))
-                .when()
-                .post(USER_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        def actualData = response.jsonPath().get(TestConstants.DATA) as List<String>
-        def sortedActualData = actualData.sort()
-        def sortedExpectedData = expectedData.sort()
-        assert sortedActualData == sortedExpectedData
-
-        where:
-        uidcid   | nickName       | expectedMsg                              | expectedData
-        ""       | ""             | "Request parameter verification error: " | ["nickName must not be blank!", "uidcid must not be blank!"]
-        ""       | "name"         | "Request parameter verification error: " | ["uidcid must not be blank!"]
-        "id"     | ""             | "Request parameter verification error: " | ["nickName must not be blank!"]
-    }
-
-    def "JSON API handles invalid user updating requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload(CREATE_UPDATE_USER_JSON), uidcid, nickName))
-                .when()
-                .put(USER_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        def actualData = response.jsonPath().get(TestConstants.DATA) as List<String>
-        def sortedActualData = actualData.sort()
-        def sortedExpectedData = expectedData.sort()
-        assert sortedActualData == sortedExpectedData
-
-        where:
-        uidcid   | nickName       | expectedMsg                              | expectedData
-        ""       | ""             | "Request parameter verification error: " | ["nickName must not be blank!", "uidcid must not be blank!"]
-        ""       | "name"         | "Request parameter verification error: " | ["uidcid must not be blank!"]
-        "id"     | ""             | "Request parameter verification error: " | ["nickName must not be blank!"]
-    }
-
-    def "JSON API handles invalid user deleting requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body("[]")
-                .when()
-                .delete(USER_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        Assertions.assertEquals("deleteUser.uidcids: uidcids must not be empty!", response.jsonPath().get("msg"))
-    }
-
-    def "JSON API handles invalid graph and node creation requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload("create-graph-nodes-to-valid.json"), uidcid, title, description,
-                        temporaryId1, temporaryId2, fromId, relationName, toId))
-                .when()
-                .post(NODE_ENDPOINT + "/graph")
-                .then()
-                .extract()
-                .response()
-
-        def actualData = response.jsonPath().get(TestConstants.DATA) as List<String>
-        def sortedActualData = actualData.sort()
-        def sortedExpectedData = expectedData.sort()
-        assert sortedActualData == sortedExpectedData
-
-        where:
-        uidcid   | title      | description   | temporaryId1 | temporaryId2 | fromId | relationName | toId  | expectedMsg                              | expectedData
-        ""       | "title"    | "description" | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!"]
-        "uidcid" | ""         | "description" | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["title must not be blank!"]
-        "uidcid" | "title"    | ""            | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["description must not be blank!"]
-        "uidcid" | "title"    | "description" | ""           | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["temporaryId must not null!"]
-        "uidcid" | "title"    | "description" | "id1"        | ""           | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["temporaryId must not null!"]
-        "uidcid" | "title"    | "description" | "id1"        | "id2"        | ""     | "relation"   | "id2" | "Request parameter verification error: " | ["fromId must not be blank!"]
-        "uidcid" | "title"    | "description" | "id1"        | "id2"        | "id1"  | ""           | "id2" | "Request parameter verification error: " | ["relation must not be blank!"]
-        "uidcid" | "title"    | "description" | "id1"        | "id2"        | "id1"  | "relation"   | ""    | "Request parameter verification error: " | ["toId must not be blank!"]
-        ""       | ""         | "description" | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!", "title must not be blank!"]
-        ""       | "title"    | ""            | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!", "description must not be blank!"]
-        ""       | "title"    | "description" | ""           | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!", "temporaryId must not null!"]
-        ""       | "title"    | "description" | "id1"        | ""           | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!", "temporaryId must not null!"]
-        ""       | "title"    | "description" | "id1"        | "id2"        | ""     | "relation"   | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!", "fromId must not be blank!"]
-        ""       | "title"    | "description" | "id1"        | "id2"        | "id1"  | ""           | "id2" | "Request parameter verification error: " | ["uidcid must not be blank!", "relation must not be blank!"]
-        ""       | "title"    | "description" | "id1"        | "id2"        | "id1"  | "relation"   | ""    | "Request parameter verification error: " | ["uidcid must not be blank!", "toId must not be blank!"]
-        "uidcid" | ""         | ""            | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["title must not be blank!", "description must not be blank!"]
-        "uidcid" | ""         | "description" | ""           | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["title must not be blank!", "temporaryId must not null!"]
-        "uidcid" | ""         | "description" | "id1"        | ""           | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["title must not be blank!", "temporaryId must not null!"]
-        "uidcid" | ""         | "description" | "id1"        | "id2"        | ""     | "relation"   | "id2" | "Request parameter verification error: " | ["title must not be blank!", "fromId must not be blank!"]
-        "uidcid" | ""         | "description" | "id1"        | "id2"        | "id1"  | ""           | "id2" | "Request parameter verification error: " | ["title must not be blank!", "relation must not be blank!"]
-        "uidcid" | ""         | "description" | "id1"        | "id2"        | "id1"  | "relation"   | ""    | "Request parameter verification error: " | ["title must not be blank!", "toId must not be blank!"]
-        "uidcid" | "title"    | ""            | ""           | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["description must not be blank!", "temporaryId must not null!"]
-        "uidcid" | "title"    | ""            | "id1"        | ""           | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["description must not be blank!", "temporaryId must not null!"]
-        "uidcid" | "title"    | ""            | "id1"        | "id2"        | ""     | "relation"   | "id2" | "Request parameter verification error: " | ["description must not be blank!", "fromId must not be blank!"]
-        "uidcid" | "title"    | ""            | "id1"        | "id2"        | "id1"  | ""           | "id2" | "Request parameter verification error: " | ["description must not be blank!", "relation must not be blank!"]
-        "uidcid" | "title"    | ""            | "id1"        | "id2"        | "id1"  | "relation"   | ""    | "Request parameter verification error: " | ["description must not be blank!", "toId must not be blank!"]
-        "uidcid" | "title"    | "description" | ""           | ""           | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["temporaryId must not null!", "temporaryId must not null!"]
-        "uidcid" | "title"    | "description" | ""           | "id2"        | ""     | "relation"   | "id2" | "Request parameter verification error: " | ["temporaryId must not null!", "fromId must not be blank!"]
-    }
-
-    def "JSON API handles invalid graph retrieving requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload(GET_GRAPH_JSON), ""))
-                .when()
-                .post(GRAPH_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        Assertions.assertEquals("Request parameter verification error: ", response.jsonPath().get("msg"))
-        Assertions.assertEquals("uuid must not be blank!", response.jsonPath().get("data[0]"))
-    }
-
-    def "JSON API handles invalid graph updating requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload(UPDATE_GRAPH_JSON), "", TestConstants.TEST_TILE1))
-                .when()
-                .put(GRAPH_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        Assertions.assertEquals("Request parameter verification error: ", response.jsonPath().get("msg"))
-        Assertions.assertEquals("uuid must not be blank!", response.jsonPath().get("data[0]"))
-    }
-
-    def "JSON API handles invalid graph deleting requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload("delete-graph-to-valid.json"), uidcid))
-                .when()
-                .delete(GRAPH_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        def actualData = response.jsonPath().get(TestConstants.DATA) as List<String>
-        def sortedActualData = actualData.sort()
-        def sortedExpectedData = expectedData.sort()
-        assert sortedActualData == sortedExpectedData
-
-        where:
-        uidcid   | expectedMsg                              | expectedData
-        ""       | "Request parameter verification error: " | ["uidcid must not be blank!", "uuids must not be empty!"]
-        "id"     | "Request parameter verification error: " | ["uuids must not be empty!"]
-    }
-
-    def "JSON API handles invalid node creation requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload("create-nodes-to-valid.json"), graphUuid,
-                        temporaryId1, temporaryId2, fromId, relationName, toId))
-                .when()
-                .post(NODE_ENDPOINT)
-                .then()
-                .extract()
-                .response()
-
-        def actualData = response.jsonPath().get(TestConstants.DATA) as List<String>
-        def sortedActualData = actualData.sort()
-        def sortedExpectedData = expectedData.sort()
-        assert sortedActualData == sortedExpectedData
-
-        where:
-        graphUuid | temporaryId1 | temporaryId2 | fromId | relationName | toId  | expectedMsg                              | expectedData
-        ""        | "id1"        | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uuid must not be blank!"]
-        "id"      | ""           | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["temporaryId must not null!"]
-        ""        | ""           | "id2"        | "id1"  | "relation"   | "id2" | "Request parameter verification error: " | ["uuid must not be blank!", "temporaryId must not null!"]
-        "id"      | "id1"        | "id2"        | ""     | "relation"   | "id2" | "Request parameter verification error: " | ["fromId must not be blank!"]
-        "id"      | "id1"        | "id2"        | "id1"  | "relation"   | ""    | "Request parameter verification error: " | ["toId must not be blank!"]
-        "id"      | "id1"        | "id2"        | ""     | "relation"   | ""    | "Request parameter verification error: " | ["fromId must not be blank!", "toId must not be blank!"]
-    }
-
-    def "JSON API handles invalid node binding requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload("relate-node-to-valid.json"), fromId, relationName, toId))
-                .when()
-                .post(NODE_ENDPOINT + "/bind")
-                .then()
-                .extract()
-                .response()
-
-        def actualMsg = response.jsonPath().get("msg")
-        assert actualMsg == expectedMsg
-
-        where:
-        fromId | relationName | toId  | expectedMsg
-        ""     | "relation"   | "id2" | "bindNodes.dtos[0].fromId: uuid must not be blank!"
-        "id1"  | ""           | "id2" | "bindNodes.dtos[0].relationName: relation must not be blank!"
-        "id1"  | "relation"   | ""    | "bindNodes.dtos[0].toId: uuid must not be blank!"
-    }
-
-    def "JSON API handles invalid node updating requests"() {
-        expect:
-        Response response = RestAssured
-                .given()
-                .contentType(ContentType.JSON)
-                .accept(ContentType.JSON)
-                .body(String.format(payload(UPDATE_NODE_JSON), "", ""))
-                .when()
-                .post(NODE_ENDPOINT + "/update")
-                .then()
-                .extract()
-                .response()
-
-        Assertions.assertEquals("Request parameter verification error: ", response.jsonPath().get("msg"))
-        Assertions.assertEquals("uuid must not be blank!", response.jsonPath().get("data[0]"))
-    }
-
     def "JSON API allows for POSTing, GETing, PUTTing, and DELETing the user, graph and node"() {
         expect: "database is initially empty"
         Response getUserResponse = RestAssured
@@ -360,7 +133,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .get(USER_ENDPOINT)
         getUserResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals([], getUserResponse.jsonPath().getList(TestConstants.DATA))
 
@@ -377,7 +150,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         postUserResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         Assertions.assertEquals(postUserResponse.jsonPath().get("data.uidcid"), TEST_UIDCID)
         Assertions.assertEquals(postUserResponse.jsonPath().get("data.nickName"), TEST_NICK_NAME)
@@ -406,7 +179,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .put(USER_ENDPOINT)
                 .then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "we can GET that User entity with updated attribute"
         Response getUpdatedUserEntityResponse = RestAssured
@@ -418,7 +191,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         getUpdatedUserEntityResponse.then()
-                .statusCode(200)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(TEST_UIDCID, getUpdatedUserEntityResponse.jsonPath().get("data.uidcid"))
         Assert.assertEquals(UPDATE_NICK_NAME, getUpdatedUserEntityResponse.jsonPath().get("data.nickName"))
@@ -436,7 +209,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         postGraphResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         Assertions.assertNotNull(postGraphResponse.jsonPath().get("data.uuid"))
         Assertions.assertEquals(postGraphResponse.jsonPath().get("data.title"), TEST_GRAPH_TITLE)
@@ -451,7 +224,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         getGraphEntityResponse.then()
-                .statusCode(200)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(TEST_UIDCID, getGraphEntityResponse.jsonPath().get("data.uidcid"))
         Assert.assertEquals(TEST_GRAPH_TITLE, getGraphEntityResponse.jsonPath().get("data.graphs[0].title"))
@@ -466,7 +239,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .put(GRAPH_ENDPOINT)
                 .then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "we can GET that Graph entity with updated attribute"
         Response getUpdatedGraphEntityResponse = RestAssured
@@ -478,7 +251,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         getUpdatedGraphEntityResponse.then()
-                .statusCode(200)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(getGraphEntityResponse.jsonPath().get("data.graphs[0].uuid"),
                 getUpdatedGraphEntityResponse.jsonPath().get("data.graphs[0].uuid"))
@@ -499,7 +272,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         postNodeResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "we can GET that Node entity next"
         Response getNodeEntityResponse = RestAssured
@@ -531,7 +304,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .post(NODE_ENDPOINT + "/update")
                 .then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "we can GET that Node entity with updated attribute"
         Response getUpdatedNodeEntityResponse = RestAssured
@@ -543,7 +316,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         getUpdatedNodeEntityResponse.then()
-                .statusCode(200)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(getNodeEntityResponse.jsonPath().get("data.nodes[0].uuid"),
                 getUpdatedNodeEntityResponse.jsonPath().get("data.uuid"))
@@ -593,7 +366,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .delete(NODE_ENDPOINT)
         deleteNodeResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "that Node entity is not found in database anymore"
         Response getResponse4 = RestAssured
@@ -601,7 +374,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .get(NODE_ENDPOINT + "/" + getNodeEntityResponse.jsonPath().get("data.nodes[1].startNode.uuid"))
         getResponse4.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(null, getResponse4.jsonPath().get(TestConstants.DATA))
 
@@ -620,7 +393,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         postGraphResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         Assertions.assertNotNull(postGraphResponse.jsonPath().get("data.uuid"))
         Assertions.assertEquals(postGraphResponse.jsonPath().get("data.title"), TEST_GRAPH_TITLE)
@@ -640,7 +413,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         getGraphEntityFilterResponse.then()
-                .statusCode(200)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(TestConstants.DEFALUT_PAGE_NUMBER, getGraphEntityFilterResponse
                 .jsonPath().get("data.pageNumber"))
@@ -671,7 +444,7 @@ abstract class AbstractITSpec extends Specification {
                 .response()
 
         getGraphEntityFilterResponsePage1.then()
-                .statusCode(200)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(TestConstants.DEFALUT_PAGE_NUMBER, getGraphEntityFilterResponsePage1
                 .jsonPath().get("data.pageNumber"))
@@ -713,7 +486,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .delete(GRAPH_ENDPOINT)
         deleteGraphResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "that Graph entity is not found in database anymore"
         Response getResponse3 = RestAssured
@@ -739,7 +512,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .delete(USER_ENDPOINT)
         deleteUserResponse.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         then: "that User entity is not found in database anymore"
         Response getResponse2 = RestAssured
@@ -747,7 +520,7 @@ abstract class AbstractITSpec extends Specification {
                 .when()
                 .get(USER_ENDPOINT + "/" + TEST_UIDCID)
         getResponse2.then()
-                .statusCode(OK_CODE)
+                .statusCode(HttpStatus.OK.value())
 
         Assert.assertEquals(null, getResponse2.jsonPath().getList(TestConstants.DATA))
     }
